requirements.txt
Pyrogram==1.2.9
TgCrypto==1.2.2
ffmpeg-python
psutil
pytgcalls==0.0.17
wheel
tgintegration
youtube-dl
youtube_search
requests
